// ehif-dbi/src/main/frontend/src/app/api/models/error.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { NodePath } from '../models/node-path';
import { SchemaLocation } from '../models/schema-location';
export interface Error {
  arguments?: Array<{
}>;
  details?: {
[key: string]: {
};
};
  evaluationPath?: NodePath;
  index?: number;
  instanceLocation?: NodePath;
  keyword?: string;
  message?: string;
  messageKey?: string;
  property?: string;
  schemaLocation?: SchemaLocation;
}